Add drawing all pieces in starting positions

+++ chess/constants.py
@@ -2,6 +2,6 @@ class UiConstants():
     windowSize = 600
 
     BOARD_PERCENT_OF_WINDOW = 0.9
-    boardSize = windowSize * BOARD_PERCENT_OF_WINDOW
+    boardSize = int(windowSize * BOARD_PERCENT_OF_WINDOW)
 
-    squareSize = boardSize / 8
+    squareSize = int(boardSize / 8)
